test(c): merge status + history cores -> integration/test_readonly_commands.py (#40)
Seven tests (3 from test_status_core, 4 from test_history_core) merged
into one file since both are thin orchestrators over StateManager.
Classes flattened per Rule M2; test names disambiguated where original
names collided.
